TL: refactored to avoid merge conflicts and other collisions, modularity etc. Some cleanup.
