+++ src/asl2txt.py
@@ -1,0 +1,7 @@
+import gui
+
+def main():
+   gui.go()
+
+if __name__ == '__main__':
+    main()
