+++ src/distalData.py
@@ -1,0 +1,205 @@
+#letter_angles = [thumb and thumb,
+#   thumb and index,
+#   thumb and middle,
+#   thumb and ring,
+#   thumb and pinkly,
+#   index and thumb,
+#   index and index,
+#   index and middle,
+#   index and ring,
+#   index and pinky,
+#   middle and thumb,
+#   middle and index,
+#   middle and middle,
+#   middle and ring,
+#   middle and pinky,
+#   ring and thumb,
+#   ring and index,
+#   ring and middle,
+#   ring and ring,
+#   ring and pinky,
+#   pinky and thumb,
+#   pinky and index,
+#   pinky and middle,
+#   pinky and ring,
+#   pinky and pinky]
+
+A_angles = [0.0,
+  2.8812789917,
+  2.859770298,
+  2.84813141823,
+  2.73459911346,
+  2.8812789917,
+  0.0,
+  0.118831209838,
+  0.228779360652,
+  0.383088201284,
+  2.859770298,
+  0.118831209838,
+  0.0,
+  0.117045581341,
+  0.267506659031,
+  2.84813141823,
+  0.228779360652,
+  0.117045581341,
+  0.0,
+  0.154956310987,
+  2.73459911346,
+  0.383088201284,
+  0.267506659031,
+  0.154956310987,
+  0.0]
+B_angles =[0.0,
+   0.439740777016,
+   0.527846693993,
+   0.577073693275,
+   0.63708627224,
+   0.439740777016,
+   0.0,
+   0.100545287132,
+   0.154377490282,
+   0.201784655452,
+   0.527846693993,
+   0.100545287132,
+   0.0,
+   0.0538404807448,
+   0.109854474664,
+   0.577073693275,
+   0.154377490282,
+   0.0538404807448,
+   0.0,
+   0.0698691532016,
+   0.63708627224,
+   0.201784655452,
+   0.109854474664,
+   0.0698691532016,
+   0.0]
+C_angles = [0.0,
+ 0.565682411194,
+ 0.548766136169,
+ 0.647555172443,
+ 1.09946036339,
+ 0.565682411194,
+ 0.0,
+ 0.0179408397526,
+ 0.352951109409,
+ 0.841872930527,
+ 0.548766136169,
+ 0.0179408397526,
+ 0.0,
+ 0.346947342157,
+ 0.840921521187,
+ 0.647555172443,
+ 0.352951109409,
+ 0.346947342157,
+ 0.0,
+ 0.51339751482,
+ 1.09946036339,
+ 0.841872930527,
+ 0.840921521187,
+ 0.51339751482,
+ 0.51339751482]
+D_angles = [0.0,
+   0.880243360996,
+   1.00771903992,
+   1.9491314888,
+   2.23631477356,
+   0.880243360996,
+   0.0,
+   0.395872354507,
+   2.08581185341,
+   2.18799972534,
+   1.00771903992,
+   0.395872354507,
+   0.0,
+   1.69160568714,
+   1.80022823811,
+   1.9491314888,
+   2.08581185341,
+   1.69160568714,
+   0.0,
+   0.291275531054,
+   2.23631477356,
+   2.18799972534,
+   1.80022823811,
+   0.291275531054,
+   0.291275531054]
+E_angles = [ 0.0,
+ 0.308618754148,
+ 2.73403167725,
+ 2.82854390144,
+ 2.84681677818,
+ 0.308618754148,
+ 0.0,
+ 3.0257191658,
+ 3.04934453964,
+ 2.96922564507,
+ 2.73403167725,
+ 3.0257191658,
+ 0.0,
+ 0.104628279805,
+ 0.17910091579,
+ 2.82854390144,
+ 3.04934453964,
+ 0.104628279805,
+ 0.0,
+ 0.0865492895246,
+ 2.84681677818,
+ 2.96922564507,
+ 0.17910091579,
+ 0.0865492895246,
+ 0.0865492895246]
+F_angles = [ 0.0,
+ 1.96462583542,
+ 0.490814357996,
+ 0.490148037672,
+ 0.459969967604,
+ 1.96462583542,
+ 0.0,
+ 2.21380853653,
+ 2.40438628197,
+ 2.42444562912,
+ 0.490814357996,
+ 2.21380853653,
+ 0.0,
+ 0.249927476048,
+ 0.422325372696,
+ 0.490148037672,
+ 2.40438628197,
+ 0.249927476048,
+ 0.0,
+ 0.190423130989,
+ 0.459969967604,
+ 2.42444562912,
+ 0.422325372696,
+ 0.190423130989,
+ 0.190423130989]
+G_angles = [0.0,
+ 1.06423842907,
+ 0.987863242626,
+ 1.54763281345,
+ 2.09381890297,
+ 1.06423842907,
+ 0.0,
+ 0.189559593797,
+ 1.38221013546,
+ 1.9089962244,
+ 0.987863242626,
+ 0.189559593797,
+ 0.0,
+ 1.22131979465,
+ 1.78623831272,
+ 1.54763281345,
+ 1.38221013546,
+ 1.22131979465,
+ 0.0,
+ 0.626338541508,
+ 2.09381890297,
+ 1.9089962244,
+ 1.78623831272,
+ 0.626338541508,
+ 0.626338541508]
+all_letter_angles = [A_angles, B_angles, C_angles, D_angles, E_angles, F_angles, G_angles]
+letter_names = [
+ "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M",
+ "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z"]
